pkg/util: relocate single-consumer sub-packages closer to their consumers

pkg/util: relocate single-consumer sub-packages closer to their consumers
