+++ cmd/cniplugins/passt-binding/pkg/plugin/sysctl/sysctl.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strconv"
 
-	"kubevirt.io/kubevirt/pkg/util/sysctl"
+	"kubevirt.io/kubevirt/pkg/network/sysctl"
 )
 
 type sysControl struct{}
