+++ pkg/network/driver/procsys/BUILD.bazel
@@ -8,5 +8,5 @@ go_library(
     ],
     importpath = "kubevirt.io/kubevirt/pkg/network/driver/procsys",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/util/sysctl:go_default_library"],
+    deps = ["//pkg/network/sysctl:go_default_library"],
 )
