+++ pkg/network/sysctl/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["sysctl.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/util/sysctl",
+    importpath = "kubevirt.io/kubevirt/pkg/network/sysctl",
     visibility = ["//visibility:public"],
     deps = ["//pkg/util:go_default_library"],
 )
