+++ pkg/virt-controller/watch/migration/BUILD.bazel
@@ -19,7 +19,7 @@ go_library(
         "//pkg/tpm:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/migrations:go_default_library",
-        "//pkg/util/trace:go_default_library",
+        "//pkg/virt-controller/watch/trace:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//staging/src/kubevirt.io/api/backup/v1alpha1:go_default_library",
