+++ pkg/virt-controller/watch/pool/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
         "//pkg/apimachinery/patch:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/pointer:go_default_library",
-        "//pkg/util/trace:go_default_library",
+        "//pkg/virt-controller/watch/trace:go_default_library",
         "//pkg/virt-controller/watch/common:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/api/pool/v1beta1:go_default_library",
