+++ pkg/virt-controller/watch/topology/BUILD.bazel
@@ -14,7 +14,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/pointer:go_default_library",
-        "//pkg/util/nodes:go_default_library",
+        "//pkg/virt-controller/watch/topology/nodes:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
