+++ pkg/virt-controller/watch/trace/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["trace.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/util/trace",
+    importpath = "kubevirt.io/kubevirt/pkg/virt-controller/watch/trace",
     visibility = ["//visibility:public"],
     deps = ["//vendor/k8s.io/utils/trace:go_default_library"],
 )
