+++ pkg/virt-controller/watch/vm/BUILD.bazel
@@ -26,7 +26,7 @@ go_library(
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/util/migrations:go_default_library",
-        "//pkg/util/trace:go_default_library",
+        "//pkg/virt-controller/watch/trace:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/watch/common:go_default_library",
         "//pkg/virt-controller/watch/descheduler:go_default_library",
