+++ pkg/virt-controller/watch/vmi/BUILD.bazel
@@ -21,7 +21,7 @@ go_library(
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
         "//pkg/util/migrations:go_default_library",
-        "//pkg/util/trace:go_default_library",
+        "//pkg/virt-controller/watch/trace:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/common:go_default_library",
